the concept of using queues

+++ app/Http/Controllers/JobController.php
@@ -1,14 +1,20 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Models\Job;
 
+use App\Mail\JobPosted;
+use App\Models\Job;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
     public function index()
     {
         $jobs = Job::with('employer')->latest()->simplePaginate(3);
+
         return view('jobs.index', [
             'jobs' => $jobs
         ]);
@@ -21,48 +27,55 @@ class JobController extends Controller
 
     public function show(Job $job)
     {
-        return view('jobs.show',['job'=>$job]);
+        return view('jobs.show', ['job' => $job]);
     }
 
     public function store()
     {
         request()->validate([
-        'title'=>['required','min:3'],
-        'salary'=>['required'],
-    ]);
+            'title' => ['required', 'min:3'],
+            'salary' => ['required']
+        ]);
 
-    Job::create([
-        'title'=>request('title'),
-        'salary'=>request('salary'),
-        'employer_id'=> 1
-    ]);
-    return redirect('/jobs');
+        $job=Job::create([
+            'title' => request('title'),
+            'salary' => request('salary'),
+            'employer_id' => 1
+        ]);
 
+        Mail::to($job->employer->user)->queue(new JobPosted($job));
+
+        return redirect('/jobs');
     }
 
     public function edit(Job $job)
     {
-        return view('jobs.edit',['job'=>$job]);
+        return view('jobs.edit', ['job' => $job]);
     }
 
     public function update(Job $job)
     {
-        request()->validate([
-            'title'=>['required','min:3'],
-            'salary'=>['required'],
-        ]);
-        $job->update([
-            'title'=>request('title'),
-            'salary'=>request('salary'),
-        ]);
-       
-        return redirect("/jobs/{$job->id}");
+        Gate::authorize('edit-job', $job);
 
+        request()->validate([
+            'title' => ['required', 'min:3'],
+            'salary' => ['required']
+        ]);
+
+        $job->update([
+            'title' => request('title'),
+            'salary' => request('salary'),
+        ]);
+
+        return redirect('/jobs/' . $job->id);
     }
 
     public function destroy(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
+
         return redirect('/jobs');
     }
 }
